//リファクタリングの途中（2021/7/30）

+++ GameTemplate/Game/SpriteRender.cpp
@@ -9,8 +9,8 @@ bool SpriteRender::Start() {
 void SpriteRender::Init(const char* filepath, float width, float height, AlphaBlendMode mode) {
 
 	m_initData.m_ddsFilePath[enData_Zeroth] = filepath;
-	m_initData.m_width = width;
-	m_initData.m_height = height;
+	m_initData.m_width = static_cast<UINT>(width);
+	m_initData.m_height = static_cast<UINT>(height);
 	m_initData.m_fxFilePath = "Assets/shader/sprite.fx";
 	m_initData.m_alphaBlendMode = mode;
